Add stopping container via CLI, refactor directory structure
