+++ main.py
@@ -3,7 +3,7 @@ An entry point of the client application that
 a user should run.
 """
 
-from cli import app
+from app.cli import app
 
 if __name__ == '__main__':
     app()
